moved Navigator.tsx NavigatorExtension.tsx navigatorComponents.tsx to ui/components/layout/Navigator/
