+++ ui/components/layout/Navigator/NavigatorExtension.tsx
@@ -4,38 +4,38 @@ import {
   getDependencies,
   createRequires,
 } from '@paciolan/remote-component';
-import MesheryPerformanceComponent from './Performance';
-import PatternServiceFormCore from './MesheryMeshInterface/PatternServiceFormCore';
-import InfoModal from '../components/shared/Modal/Information/InfoModal';
-import ConfigurationSubscription from '../components/graphql/subscriptions/ConfigurationSubscription';
-import _PromptComponent from './PromptComponent';
-import { CapabilitiesRegistry } from '../utils/disabledComponents';
-import { useNotification } from '../utils/hooks/useNotification';
-import Modal from './shared/Modal/Modal';
-import ExportModal from './shared/Modal/ExportModal';
-import { FormatStructuredData } from './DataFormatter';
+import MesheryPerformanceComponent from '../../Performance';
+import PatternServiceFormCore from '../../MesheryMeshInterface/PatternServiceFormCore';
+import InfoModal from '../../shared/Modal/Information/InfoModal';
+import ConfigurationSubscription from '@/graphql/subscriptions/ConfigurationSubscription';
+import _PromptComponent from '../../PromptComponent';
+import { CapabilitiesRegistry } from '../../../utils/disabledComponents';
+import { useNotification } from '../../../utils/hooks/useNotification';
+import Modal from '../../shared/Modal/Modal';
+import ExportModal from '../../shared/Modal/ExportModal';
+import { FormatStructuredData } from '../../DataFormatter';
 import { useFilterK8sContexts } from '@/utils/hooks/useKubernetesHook';
 import { useDynamicComponent } from '@/utils/context/dynamicContext';
-import { ValidateDesign } from './designs/lifecycle/ValidateDesign';
-import { DryRunDesign } from './designs/lifecycle/DryRun';
-import { DeployStepper, UnDeployStepper } from './designs/lifecycle/DeployStepper';
+import { ValidateDesign } from '../../DesignLifeCycle/ValidateDesign';
+import { DryRunDesign } from '../../DesignLifeCycle/DryRun';
+import { DeployStepper, UnDeployStepper } from '../../DesignLifeCycle/DeployStepper';
 import { designValidationMachine } from 'machines/validator/designValidator';
 import CAN from '@/utils/can';
 import { mesheryEventBus } from '@/utils/eventBus';
 import { ThemeTogglerCore } from '@/themes/hooks';
-import RJSFForm from './MesheryMeshInterface/PatternService/RJSF';
-import { DynamicFullScreenLoader } from './LoadingComponents/DynamicFullscreenLoader';
-import Troubleshoot from './TroubleshootingComponent';
-import TypingFilter from './TypingFilter';
-import CreateModelModal from './Settings/Registry/CreateModelModal';
-import ImportModelModal from './Settings/Registry/ImportModelModal';
-import { ViewInfoModal } from './shared/Modal/ViewInfoModal';
+import RJSFForm from '../../MesheryMeshInterface/PatternService/RJSF';
+import { DynamicFullScreenLoader } from '../../LoadingComponents/DynamicFullscreenLoader';
+import Troubleshoot from '../../TroubleshootingComponent';
+import TypingFilter from '../../TypingFilter';
+import CreateModelModal from '../../Settings/Registry/CreateModelModal';
+import ImportModelModal from '../../Settings/Registry/ImportModelModal';
+import { ViewInfoModal } from '../../shared/Modal/ViewInfoModal';
 import { selectK8sConfig, selectSelectedK8sClusters } from '@/store/slices/mesheryUi';
 import { useSelector } from 'react-redux';
-import { store } from '../store';
+import { store } from '../../../store';
 import ProviderStoreWrapper from '@/store/ProviderStoreWrapper';
 import { WorkspaceModalContext } from '@/utils/context/WorkspaceModalContextProvider';
-import { RelationshipEvaluationTraceFormatter } from './NotificationCenter/formatters/relationship_evaluation';
+import { RelationshipEvaluationTraceFormatter } from '../../NotificationCenter/formatters/relationship_evaluation';
 import { useRegistryModal } from '@/utils/hooks/useRegistryModal';
 
 const requires = createRequires(getDependencies);
